+++ unlock-js/src/walletService.js
@@ -4,6 +4,7 @@ import FetchJsonProvider from './FetchJsonProvider'
 import { GAS_AMOUNTS } from './constants'
 import utils from './utils'
 import { generateKeyMetadataPayload } from './typedData/keyMetadata'
+import { generateKeyHolderMetadataPayload } from './typedData/keyHolderMetadata'
 import 'cross-fetch/polyfill'
 
 const bytecode = require('./bytecode').default
@@ -350,13 +351,17 @@ export default class WalletService extends UnlockService {
   /**
    * Sign and send a request to update metadata specific to a given key.
    *
-   * @param {string} lockAddress - The address of the lock
-   * @param {string} keyId - The id of the key to set metadata on
-   * @param {Object.<string, string>} metadata - The metadata fields and values to set
-   * @param {string} locksmithHost - A url with no trailing slash
+   * @param {Object} params
+   * @param {string} params.lockAddress - The address of the lock
+   * @param {string} params.keyId - The id of the key to set metadata on
+   * @param {Object.<string, string>} params.metadata - The metadata fields and values to set
+   * @param {string} params.locksmithHost - A url with no trailing slash
    * @param {*} callback
    */
-  async setKeyMetadata(lockAddress, keyId, metadata, locksmithHost, callback) {
+  async setKeyMetadata(
+    { lockAddress, keyId, metadata, locksmithHost },
+    callback
+  ) {
     const url = `${locksmithHost}/api/key/${lockAddress}/${keyId}`
     try {
       const currentAddress = await this.getAccount()
@@ -390,20 +395,93 @@ export default class WalletService extends UnlockService {
   }
 
   /**
-   * Sign and send a request to read metadata specific to a given key.
+   * @typedef {Object} keyholderMetadata
+   * @property {Object.<string, string>} [publicData={}] - Publicly available metadata
+   * @property {Object.<string, string>} [protectedData={}] - Restricted access metadata
+   */
+
+  /**
+   * Sign and send a request to update metadata specific to a given
+   * user address for a given lock.
    *
-   * @param {string} lockAddress - The address of the lock
-   * @param {string} keyId - The id of the key to set metadata on
-   * @param {string} locksmithHost - A url with no trailing slash
+   * @param {Object} params
+   * @param {string} params.lockAddress - The address of the lock
+   * @param {string} params.userAddress - The address of the user (optional, defaults to current wallet user)
+   * @param {keyholderMetadata} params.metadata - The metadata fields and values to set
+   * @param {string} params.locksmithHost - A url with no trailing slash
    * @param {*} callback
    */
-  async getKeyMetadata(lockAddress, keyId, locksmithHost, callback) {
+  async setUserMetadata(
+    { lockAddress, userAddress, metadata, locksmithHost },
+    callback
+  ) {
+    const url = `${locksmithHost}/api/key/${lockAddress}/user/${userAddress}`
+    try {
+      const currentAddress = await this.getAccount()
+      const payload = generateKeyHolderMetadataPayload(currentAddress, metadata)
+      const signature = await this.unformattedSignTypedData(
+        currentAddress,
+        payload
+      )
+
+      const response = await fetch(url, {
+        method: 'PUT',
+        headers: {
+          Authorization: `Bearer ${Buffer.from(signature).toString('base64')}`,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(payload),
+      })
+
+      if (response.status !== 202) {
+        callback(
+          new Error(
+            `Received ${response.status} from locksmith: ${response.statusText}`
+          )
+        )
+        return
+      }
+      callback(null, true)
+    } catch (error) {
+      callback(error, null)
+    }
+  }
+
+  /**
+   * Sign and send a request to read metadata specific to a given key.
+   *
+   * @param {Object} params
+   * @param {string} params.lockAddress - The address of the lock
+   * @param {string} params.keyId - The id of the key to read metadata on
+   * @param {string} params.locksmithHost - A url with no trailing slash
+   * @param {boolean} params.getProtectedData - when truthy, will generate signature to get protected metadata
+   * @param {*} callback
+   */
+  async getKeyMetadata(
+    { lockAddress, keyId, locksmithHost, getProtectedData },
+    callback
+  ) {
     const url = `${locksmithHost}/api/key/${lockAddress}/${keyId}`
     try {
-      const response = await fetch(url, {
+      let options = {
         method: 'GET',
         accept: 'json',
-      })
+      }
+
+      if (getProtectedData) {
+        const currentAddress = await this.getAccount()
+        const payload = generateKeyMetadataPayload(currentAddress, {})
+        const signature = await this.unformattedSignTypedData(
+          currentAddress,
+          payload
+        )
+        options.Authorization = `Bearer ${Buffer.from(signature).toString(
+          'base64'
+        )}`
+      }
+
+      const response = await fetch(url, options)
+
       const json = await response.json()
       callback(null, json)
     } catch (error) {